Revisa javaDoc

+++ MeioDoQuadrado.java
@@ -1,10 +1,12 @@
 /**
  * Implementação da função hash utilizando o método do meio do quadrado.
  * 
- * A chave é elevada ao quadrado e uma região central dos bits do resultado é utilizada para gerar o índice da tabela hash.
- * Essa técnica tenta reduzir padrões existentes nas chaves originais e favorecer uma distribuição mais uniforme dos elementos.
+ * A chave é elevada ao quadrado e uma região central dos bits do
+ * resultado é utilizada para gerar o índice da tabela hash.
+ * Essa técnica busca reduzir padrões existentes nas chaves originais
+ * e favorecer uma distribuição mais uniforme dos elementos na tabela hash.
  */
-public class MeioDoQuadrado extends FuncaoHash{
+public class MeioDoQuadrado extends FuncaoHash {
     public MeioDoQuadrado(int size){
 
         super(size);
@@ -12,21 +14,23 @@ public class MeioDoQuadrado extends FuncaoHash{
     }
 
     /**
-     * Calcula o índice hash para uma determinada chave.
+     * Calcula o índice hash da chave utilizando o método do meio do quadrado.
      * 
      * @param chave A chave a ser transformada
      * @return índice válido na tabela hash
      */
     @Override
     protected int hash(int chave) {
-        // Passo 1: Converte a chave para um valor long para evitar overflow ao calcular o quadrado.
+        // Passo 1: Converte a chave para um inteiro sem sinal de 64 bits,
+        // evitando problemas com valores negativos durante o cálculo.
         long valor = Integer.toUnsignedLong(chave);
         
         // Passo 2: Calcula o quadrado da chave.
         long quadrado = valor * valor;
 
-        // Passo 3: Extrai bits centrais do quadrado, deslocando 16 bits à direita e descartando os 16 bits menos significativos e preservando os 32 bits seguintes.
-        long meio = (int)((quadrado >>> 16) & 0xFFFFFFFFL);
+        // Passo 3: Extrai bits centrais do quadrado.
+        // O deslocamento remove os 16 bits menos significativos e preservando os 32 bits seguintes.
+        long meio = (quadrado >>> 16) & 0xFFFFFFFFL;
 
         // Passo 4: Retorna o índice hash, garantindo que esteja dentro do tamanho da tabela.
         return (int) Math.floorMod(meio, this.size);
